Moved predict method
