+++ catalax/neural/neuralbase.py
@@ -1,13 +1,16 @@
 import json
 import os
-from typing import Dict, List
-from copy import deepcopy
+from typing import Dict, List, Optional
 
 import diffrax
 import equinox as eqx
+import jax
 import jax.random as jrandom
+import jax.numpy as jnp
 
 from catalax import Model
+from catalax.tools.simulation import Stack
+
 from .mlp import MLP
 
 
@@ -16,6 +19,7 @@ class NeuralBase(eqx.Module):
     observable_indices: List[int]
     hyperparams: Dict
     solver: diffrax.AbstractSolver
+    vector_field: Optional[Stack]
 
     def __init__(
         self,
@@ -25,13 +29,14 @@ class NeuralBase(eqx.Module):
         model: Model,
         observable_indices: List[int],
         solver=diffrax.Tsit5,
-        **kwargs
+        **kwargs,
     ):
         super().__init__(**kwargs)
 
         # Save solver and observable indices
         self.solver = solver
         self.observable_indices = observable_indices
+        self.vector_field = None
 
         # Keep hyperparams for serialisation
         self.hyperparams = {
@@ -41,6 +46,39 @@ class NeuralBase(eqx.Module):
             "model": model.to_dict(),
         }
 
+    def predict(
+        self,
+        y0s: jax.Array,
+        t0: int = 0,
+        t1: Optional[int] = None,
+        nsteps: int = 1000,
+        times: Optional[jax.Array] = None,
+    ):
+        if times is None:
+            assert t1 is not None, "Either times or t1 must be given."
+
+            # Generate time points, if not explicitly given
+            times = jnp.linspace(t0, t1, nsteps)
+
+        if y0s.shape[0] > 1 and len(times.shape) == 1:
+            # If multiple initial conditions are given, repeat time points
+            times = jnp.stack([times] * y0s.shape[0], axis=0)
+
+        # Single simulation case
+        if len(times.shape) == 1 and len(y0s.shape) == 1:
+            return times, self(times, y0s)  # type: ignore
+
+        # Batch simulation case
+        assert (
+            len(times.shape) == 2 and len(y0s.shape) == 2
+        ), f"Incompatible shapes: time shape = {times.shape}, y0 shape = {y0s.shape}. Both must be 2D."
+
+        assert (
+            times.shape[0] == y0s.shape[0]
+        ), f"Incompatible shapes: time shape = {times.shape}, y0.shape = {y0s.shape}. First dimension must be equal in case of batches."
+
+        return times, jax.vmap(self, in_axes=(0, 0))(times, y0s)  # type: ignore
+
     @classmethod
     def from_model(
         cls,
@@ -49,7 +87,7 @@ class NeuralBase(eqx.Module):
         depth: int,
         key,
         solver=diffrax.Tsit5,
-        **kwargs
+        **kwargs,
     ):
         """Intializes a NeuralODE from a catalax.Model
 
@@ -72,7 +110,7 @@ class NeuralBase(eqx.Module):
             observable_indices=observable_indices,
             key=key,
             model=model,
-            **kwargs
+            **kwargs,
         )
 
     @classmethod
